- bugfix in api on long long division - bugfix in snd buffer - tremor codec no more buffers the whole file on play
git-svn-id: http://svn.code.sf.net/p/archopen/code@102 18109dc0-e51e-0410-b7de-bd5de724b689

+++ ArchOpen/trunk/tools/api_files/api_list.h
@@ -72,7 +72,7 @@ long # float_modsi3# long a, long b
 long # float_umodsi3# long a, long b
 long # float_divsi3# long a, long b
 long # float_udivsi3# long a, long b
-unsigned int # float_divdi3# unsigned int a, unsigned int b
+long long # float_divdi3# long long a, long long b
 
 float # float_addsf3 # float a, float b
 float # float_subsf3 # float a, float b
